requirements.txt
certifi==2020.12.5
protobuf==3.17.0
PyMySQL==1.0.2
python-dotenv==0.17.1
six==1.16.0
urllib3==1.26.4
